Remove old diversity script
Replaced abFilter_Diversity.slurm with the 06_Diversity, 07_AlphaStats,
and 07_BetaStats. The smaller scripts will be easier to deal with
when errors occur. The config/input requirements for the diversity
script and stat scripts are different. Stat scripts are more likely to
be run multiple times and/or adjusted while the diversity script will be
run once and generate the inputs for the stat scripts.