Update zymoQC config and syntax of filter-samples
References for zymo sequences and zymo taxonomy are now variables in
config that can be edited. Update output/input paths to reflect folder
structure. Syntax of feature-table filter-samples updated to be
compliant with SQL-Lite. Column names must be double-quoted to
be identifiers and literals must be single-quoted. Added the command
"taxa collapse" so that the sample feature table contains taxonomic
information to compare against in the evaluate-composition commands.